+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# Changelog of ActiveHashcash
+
+## [Unreleased]
+
+- Fix ActiveHashcash::Store#add? by converting stamp to a string
+
 ## 0.2.0 - 2022-08-02
 
 - Add ActiveHashcash::Store#clean to removed expired stamps