+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 
 - Increase complexity logarithmicly to slowdown brute force attacks
+- Add mountable dashboard to list latest stamps
 - Store stamps into the database instead of Redis
 - Fix ActiveHashcash::Store#add? by converting stamp to a string